stylesense/stylesense/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1
google-generativeai==0.5.2
groq==0.5.0
Pillow==10.3.0
requests==2.31.0
pydantic==2.6.4
